renamed mortoncode.hpp
